backend/requirements.txt
databases[postgresql]==0.9.0
fastapi[all]==0.114.2
geoalchemy2==0.15.2
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
sqlalchemy==2.0.35
uvicorn==0.30.6
